[FIX]Shryne Patch on MRP Produce Quantity is decreased.
For reference:
Commit: aab4eba86019ec888992531b7deaf48bef18e7e9 [aab4eba]
Parents: bf81e38f68
Author: Nicolas Martinelli <nim@odoo.com>
Date: Monday, December 16, 2019 1:36:47 PM
Committer: Nicolas Martinelli
[FIX] mrp: decrease produced quantity

- Create a product A with a BOM
- Create a MO for 3000 units of A
- Product 3000 units
- Unlock the MO and change the Consumed Materials and Finished Products
  quantity to less than 3000, e.g. 2900
- Update the Quantity To Produce to 2900

The Finished Products move lines are deleted, but the Consumed Materials
are left unchanged.

This is because decreasing the quantity triggers a `_do_unreserve` which
deletes the move lines:

https://github.com/odoo/odoo/blob/bf81e38f685abbd4632c7be5cc381dfa38a9b7b8/addons/stock/models/stock_move.py#L399

The issue is similar to df7953f3031a2d3b, so we use the same solution:
unreserve, but keep track of the quantities.

opw-2151927

closes odoo/odoo#41933

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>