[PERF] base_automation: compute once date on time condition based on a calendar
When using base.automation a timed condition using a calendar, the check
was done for each record to get the datetime with the delay using the
calendar.
Now we invert the logic instead of adding the delay to the record
in a loop datetime field, we substract from the last_run and now once
outside of the loop.
It goes from this logic:
last_run <= record_dt + delay < now
to this:
last_run - delay <= record_dt < now - delay

In a database With 7000 records to check, spread across 23 automated
action:
it went from 50s to 0.7s

closes odoo/odoo#171205

X-original-commit: 6708408c9d2f8137271809962f0d39ee38a0a24b
Related: odoo/enterprise#65741
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Stanislas Sobieski (sts) <sts@odoo.com>
(cherry picked from commit e78a933acac9d590ddad0273236d24e81ebd70e1)