[FIX] mrp: update durations on workorders once the MO is done
Steps to reproduce:

- Enable workorders in the settings
- Create a BOM for a product P with an operation:
    - OP1: Assembly line 1, duration 10 minutes
- Create and confirm an MO for 5 units of P
- Set a producing quantity of 2
- Validate the MO and create a back order

Issues:

On each of the operations of the MO:

The expected duration is the expected duration corresponding to the
full initial demand e.g. 50 minutes

Cause of the issues:

The expected duration is a computed field that is not updated once
the opeartion is done:
https://github.com/odoo/odoo/blob/45154cfdb044aea85ad72bb36348af7b77afd844/addons/mrp/models/mrp_workorder.py#L329-L333
However, finishing the operation will change the state to 'done' before
the MO is backordered and the `qty_production` updated. In addition,
the `_get_duration_expected` relied solely on the qty_production rather
than also relying on the actual 'qty_produced':
https://github.com/odoo/odoo/blob/45154cfdb044aea85ad72bb36348af7b77afd844/addons/mrp/models/mrp_workorder.py#L786

Fix:

We update the expected duration depending on the qty_produced in case
it was not manually set (which can be determined by comparing its
current value with the expected value before the workorder is set as
done)

opw-3873108
opw-4010550

Part-of: odoo/odoo#171771
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
(cherry picked from commit 9431c7acad3daa99b033c5da41f308d79b1bbce8)